+++ lib/sbmt/pact/provider/pact_config/base.rb
@@ -5,13 +5,20 @@ module Sbmt
     module Provider
       module PactConfig
         class Base
-          attr_reader :provider_name, :provider_version, :log_level, :provider_setup_server, :provider_setup_url, :provider_setup_port,
+          attr_reader :provider_name, :provider_version, :log_level, :provider_setup_server, :provider_setup_port, :pact_proxy_port,
             :consumer_branch, :consumer_version, :consumer_name, :broker_url, :broker_username, :broker_password, :verify_only
+
+          PACT_BROKER_FILTER_TYPE_NONE = PactBrokerProxyRunner::FILTER_TYPE_NONE
+          PACT_BROKER_FILTER_TYPE_GRPC = PactBrokerProxyRunner::FILTER_TYPE_GRPC
+          PACT_BROKER_FILTER_TYPE_ASYNC = PactBrokerProxyRunner::FILTER_TYPE_ASYNC
+          PACT_BROKER_FILTER_TYPE_SYNC = PactBrokerProxyRunner::FILTER_TYPE_SYNC
+          PACT_BROKER_FILTER_TYPE_HTTP = PactBrokerProxyRunner::FILTER_TYPE_HTTP
 
           def initialize(provider_name:, opts: {})
             @provider_name = provider_name
             @log_level = opts[:log_level] || :info
             @provider_setup_port = opts[:provider_setup_port] || 9001
+            @pact_proxy_port = opts[:provider_setup_port] || 9002
             @provider_version = opts[:provider_version] || ENV.fetch("PACT_PROVIDER_VERSION", "1.0.0")
             @consumer_branch = opts[:consumer_branch] || ENV.fetch("PACT_CONSUMER_BRANCH", nil)
             @consumer_version = opts[:consumer_version] || ENV.fetch("PACT_CONSUMER_VERSION", nil)
@@ -21,8 +28,35 @@ module Sbmt
             @broker_password = opts[:broker_password] || ENV.fetch("PACT_BROKER_PASSWORD", "")
             @verify_only = opts[:verify_only] || []
 
-            @provider_setup_url = opts[:provider_setup_url] || "http://localhost:#{@provider_setup_port}/setup-provider"
             @provider_setup_server = ProviderServerRunner.new(port: @provider_setup_port)
+            if @broker_url.present?
+              @pact_proxy_server = PactBrokerProxyRunner.new(
+                port: @pact_proxy_port, pact_broker_host: @broker_url, filter_type: filter_type,
+                pact_broker_user: @broker_username, pact_broker_password: @broker_password
+              )
+            end
+          end
+
+          def filter_type
+            PACT_BROKER_FILTER_TYPE_NONE
+          end
+
+          def start_servers
+            @provider_setup_server.start
+            @pact_proxy_server&.start
+          end
+
+          def stop_servers
+            @provider_setup_server.stop
+            @pact_proxy_server&.stop
+          end
+
+          def provider_setup_url
+            @provider_setup_server.state_setup_url
+          end
+
+          def pact_broker_proxy_url
+            @pact_proxy_server&.proxy_url
           end
 
           def new_provider_state(name, opts: {}, &block)
@@ -45,7 +79,7 @@ module Sbmt
           end
 
           def new_verifier
-            raise NotImplementedError, "#new_verifier should be implemented"
+            raise Sbmt::Pact::ImplementationRequired, "#new_verifier should be implemented"
           end
         end
       end