+++ lib/sbmt/pact/rspec/support/pact_provider_helpers.rb
@@ -1,13 +1,19 @@
 # frozen_string_literal: true
 
+require_relative "webmock/webmock_helpers"
+
 module SbmtPactProducerDsl
   module ClassMethods
     def grpc_pact_provider(provider, opts: {})
+      _pact_provider(:grpc, provider, opts: opts)
+    end
+
+    def _pact_provider(transport_type, provider, opts: {})
       raise "grpc_pact_provider is designed to be used with RSpec" unless defined?(::RSpec)
       raise "grpc_pact_provider has to be declared at the top level of a suite" unless top_level?
       raise "grpc_pact_provider is designed to be run once per provider so cannot be declared more than once" if defined?(@_pact_config)
 
-      pact_config_instance = Sbmt::Pact::Provider::PactConfig.new_grpc(provider_name: provider, opts: opts)
+      pact_config_instance = Sbmt::Pact::Provider::PactConfig.new(transport_type, provider_name: provider, opts: opts)
       instance_variable_set(:@_pact_config, pact_config_instance)
 
       # rubocop:disable RSpec/BeforeAfterAll
@@ -51,4 +57,14 @@ end
 RSpec.configure do |config|
   config.include SbmtPactProducerDsl, pact_entity: :provider
   config.extend SbmtPactProducerDsl::ClassMethods, pact_entity: :provider
+
+  config.around pact_entity: :provider do |example|
+    WebmockHelpers.turned_off do
+      if defined?(::VCR)
+        VCR.turned_off { example.run }
+      else
+        example.run
+      end
+    end
+  end
 end